Add Okta On-Prem connector compatibility changes
Add ServiceCapabilities field to ServiceProviderConfig struct
to allow customization of SCIM service capabilities.
Backwards compatibility is maintained by using the previous
list of capabilities if this list is empty.

Add SchemaExtensions field to ServiceProviderConfig struct
to allow augmenting the list of schemas returned by the
/ServiceProviderConfig end-point.

Add ExcludeResourceTypeMeta field to ResourceType struct
to allow for excluding the resource_type field in the response
meta object. This field breaks the Okta on-prem connector agent.
This field was named such to maintain backwards compatibility with
existing code that would not specify this.

Resource.response was changed to consult this new ResourceType field.

Add omitempty flag to meta.ResourceType's json tag